desktop icon
Added new desktop icon and change function of auto startup script (that runs on
startup and also runs from the desktop icon).

The new startup script will just activate the RDE Chromium window if it is
running, or start it from scratch if it is not.

In addition, startup script is now a link, to make it easier to replace startup
script when running auto boot programs (like the simon demo)

+++ start_client.sh
@@ -1,14 +1,27 @@
 #!/bin/sh
 
-# Cache corruption on sudden shutdown (or other cases) can cause missing file
-# or broken image links.  So start with a clean cache on every boot.
-rm -rf $HOME/.cache/chromium
+WINDOW_TITLE="Ready Set STEM - Chromium"
+APP_NAME="chromium"
 
-# A hard shutdown causes Chromium to report "didn't shut down correctly"...
-# Clean files responsible for this (Note: this method is ad-hoc, there's no
-# clearly defined way to do this).
-rm -f $HOME/.config/chromium/Default/Preferences
-rm -f $HOME/.config/chromium/SingletonLock
+pgrep "$APP_NAME" >/dev/null
+RUNNING=$?
+wmctrl -a "$WINDOW_TITLE"
+ACTIVATED=$?
 
-# Run!
-chromium --kiosk localhost
+echo $ACTIVATED
+echo $RUNNING
+# Full restart if not activated and running
+if [ $ACTIVATED -ne 0 -a $RUNNING -ne 0 ]; then
+    # Cache corruption on sudden shutdown (or other cases) can cause missing file
+    # or broken image links.  So start with a clean cache on every boot.
+    rm -rf $HOME/.cache/chromium
+
+    # A hard shutdown causes Chromium to report "didn't shut down correctly"...
+    # Clean files responsible for this (Note: this method is ad-hoc, there's no
+    # clearly defined way to do this).
+    rm -f $HOME/.config/chromium/Default/Preferences
+    rm -f $HOME/.config/chromium/SingletonLock
+
+    # Run!
+    chromium --kiosk localhost &
+fi
